[master] Moved lookupServer code into adjacent files so that functionality can… (#3049)
* Moved lookupServer code into adjacent files so that functionality can be disabled by switch and comparison with 8.2 sources is clean.

* Rename EthLookupServer class to EthRpcMethods as this better describes the functionality the class provides.

The side effect of inheritance means that the LookupServer class at runtime has all the features as if the methods were defined inline which helps for routines that rely upon addresses such as test mocks and rpc proc initialisation.

* Rename EthLookupServer class to EthRpcMethods as this better describes the functionality the class provides.

The side effect of inheritance means that the LookupServer class at runtime has all the features as if the methods were defined inline which helps for routines that rely upon addresses such as test mocks and rpc proc initialisation.

Inlcude gaurd

* Rename EthLookupServer class to EthRpcMethods as this better describes the functionality the class provides.

The side effect of inheritance means that the LookupServer class at runtime has all the features as if the methods were defined inline which helps for routines that rely upon addresses such as test mocks and rpc proc initialisation.

Inlcude gaurd

* FOrmat changes and a fix to ci-build.sh so you can run it locally

* Don't edit ci_build.sh too much just fix the bug
